requirements.txt
numpy
pandas
matplotlib
seaborn
tensorflow
opencv-python
imutils
scikit-learn
streamlit
Pillow